backend/requirements.txt
boto3 == 1.28.20
Django == 4.2.4
django-cors-headers == 4.2.0
django-storages == 1.13.2
djangorestframework == 3.14.0
djangorestframework-simplejwt == 5.2.2
gunicorn == 21.2.0
pandas==2.0.3
psycopg2 == 2.9.6
python-dotenv == 1.0.0
watchtower == 3.0.1
